Make calls to Resolve-DnsName explicitly use DnsClient module
On macOS and Linux platforms, the presence of alternative implementations
of Resolve-DnsName can cause us to enter an infinite loop of DNS lookups.
This has been corrected by clarifying that we only want to use the cmdlet
DnsClient\Resolve-DnsName, if present.

This bug may have affected Windows 7, too, but I have not tested this on
that obsolete platform.